chore: Move blog content to separate component
