+++ src/components/pages/blog/blog-content.module.scss
@@ -1,4 +1,4 @@
-.blog-content {
+.content {
   blockquote {
     color: $color-plum-600;
     border-left: 2px solid $color-plum-600;
